simplify submodule structure
